group compiler model modules
