+++ src/compiler/mod.rs
@@ -1,2 +1,3 @@
 pub(crate) mod load;
+pub(crate) mod model;
 pub(crate) mod syntax;
